Added the API Integration

+++ src/containers/WelcomeGuide/SignUpForm/Signup.tsx
@@ -1,8 +1,10 @@
+import { Snackbar } from "@material-ui/core";
 import React from "react";
+import WarningIcon from "@material-ui/icons/Warning";
 import { connect } from "react-redux";
 import Button from "../../../components/Button/Button";
-import StyledCheckbox from "../../../components/Checkbox/Checkbox";
 import Input from "../../../components/Input/Input";
+import { ErrorLayout, ErrorSpan } from "../../../components/Stepper/styles";
 import StrikedText from "../../../components/StrikedText/StrikedText";
 import useGenericState from "../../../Library/useGenericState";
 import { setFormValues, setShowWelcomeModal } from "../action";
@@ -13,6 +15,7 @@ import {
   Label,
   LabelContainer,
   Span,
+  StyledCheckBox,
   SubHeader,
 } from "./style";
 
@@ -20,14 +23,52 @@ const Signup = (props: any): JSX.Element => {
   const [state, setState] = useGenericState({
     username: "",
     password: "",
+    confirmPassword: "",
+    agree: false,
+    showSnackBar: false,
+    snackBarMessage: "",
   });
 
-  const { username, password } = state;
+  const {
+    username,
+    password,
+    agree,
+    showSnackBar,
+    snackBarMessage,
+    confirmPassword,
+  } = state;
 
   const handleSignup = () => {
+    if (username === "" || password === "" || confirmPassword === "") {
+      setState({
+        showSnackBar: true,
+        snackBarMessage: "Username and Password are Required",
+      });
+      return;
+    }
+
+    if (password !== confirmPassword) {
+      setState({
+        showSnackBar: true,
+        snackBarMessage: "Password Didn't Matched",
+      });
+      return;
+    }
     props.setFormValues({ username, password });
     props.setShowWelcomeModal(false);
     localStorage.setItem("showWelcomeDialog", "true");
+  };
+
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setState({ agree: event.target.checked });
+  };
+
+  const handleClose = (event?: React.SyntheticEvent, reason?: string) => {
+    if (reason === "clickaway") {
+      return;
+    }
+
+    setState({ showSnackBar: false });
   };
   return (
     <Container>
@@ -47,26 +88,47 @@ const Signup = (props: any): JSX.Element => {
         <Label>Password</Label>
         <Input
           placeholder="Enter Password"
+          type="password"
           icon="https://img.icons8.com/ios/452/password--v1.png"
           onChange={(e) => setState({ password: e.target.value })}
         />
       </InputContainer>
+      <InputContainer>
+        <Label>Remember Password</Label>
+        <Input
+          placeholder="Enter Password Again"
+          type="password"
+          icon="https://img.icons8.com/ios/452/password--v1.png"
+          onChange={(e) => setState({ confirmPassword: e.target.value })}
+        />
+      </InputContainer>
       <LabelContainer>
-        <StyledCheckbox />
+        <StyledCheckBox onChange={handleChange} />
         <Span>I agree to the terms and conditions</Span>
       </LabelContainer>
-      <Button type="primary" onClick={handleSignup}>
+
+      <Button disabled={!agree} type="primary" onClick={handleSignup}>
         Create Account
       </Button>
 
-      <StrikedText text="OR" />
+      {/* <StrikedText text="OR" />
 
       <Button type="outline" onClick={handleSignup}>
         Continue with Google
       </Button>
       <Button type="outline" onClick={handleSignup}>
         Continue with Apple
-      </Button>
+      </Button> */}
+      <Snackbar
+        open={showSnackBar}
+        autoHideDuration={2000}
+        onClose={handleClose}
+      >
+        <ErrorLayout>
+          <WarningIcon />
+          <ErrorSpan>{snackBarMessage}</ErrorSpan>
+        </ErrorLayout>
+      </Snackbar>
     </Container>
   );
 };
